refactor: derive types and config options from the schemas
